requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.0.4
psycopg2-binary==2.9.3
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
debugpy==1.5.1